+++ ci/test.sh
@@ -19,6 +19,14 @@ mkdir -p "${TEST_RESULTS_DIR}"
 COVERAGE_OPTIONS="generateHtmlReport\;assemblyFilters:+Improbable.Gdk.*,-Improbable.Gdk.Generated,-Improbable.Gdk.Generated.BuildSystem"
 COVERAGE_RESULTS_PATH="${PROJECT_DIR}/workers/unity/logs/coverage-results"
 
+if isMacOS; then
+    PLAYMODE_PLATFORM="StandaloneOSX"
+elif isLinux; then
+    PLAYMODE_PLATFORM="StandaloneLinux64"
+else
+    PLAYMODE_PLATFORM="StandaloneWindows64"
+fi
+
 traceStart "Testing Code Generator :gear:"
     dotnet test \
         --logger:"nunit;LogFilePath=${TEST_RESULTS_DIR}/code-gen-lib-test-results.xml" \
@@ -49,6 +57,7 @@ traceStart "Testing Unity: Editmode :writing_hand:"
             -coverageOptions "${COVERAGE_OPTIONS}" \
             -projectPath "${PROJECT_DIR}/workers/unity" \
             -runEditorTests \
+            -testCategory "Uncategorized" \
             -logfile "${PROJECT_DIR}/logs/editmode-test-run.log" \
             -editorTestsResultFile "${TEST_RESULTS_DIR}/editmode-test-results.xml" \
             "${ACCELERATOR_ARGS}"
@@ -59,12 +68,15 @@ traceStart "Testing Unity: Playmode :joystick:"
     pushd "workers/unity"
         dotnet run -p "${PROJECT_DIR}/.shared-ci/tools/RunUnity/RunUnity.csproj" -- \
             -batchmode \
+            -nographics \
             -enableCodeCoverage \
             -coverageResultsPath "${COVERAGE_RESULTS_PATH}/playground" \
             -coverageOptions "${COVERAGE_OPTIONS}" \
             -projectPath "${PROJECT_DIR}/workers/unity" \
             -runTests \
-            -testPlatform playmode \
+            -testPlatform ${PLAYMODE_PLATFORM} \
+            -buildTarget ${PLAYMODE_PLATFORM} \
+            -testCategory "Uncategorized" \
             -logfile "${PROJECT_DIR}/logs/playmode-test-run.log" \
             -testResults "${TEST_RESULTS_DIR}/playmode-test-results.xml" \
             "${ACCELERATOR_ARGS}"